Fixed webui.get()/getone() and simplified dispatch
Removed result exception since it was a twisted solution and didn't work
with web.py's error reporting. Instead, get and getone work by spawning
the web service in a new process and getting results with a
multiprocessing.Queue.

Simplified dispatch mecanism by simply passing a callable method to
Page, removing it's involvment in the result usage and exposure to webui
users. It now calls parser.parse_args() to get the parsed namespace and
then calls it's the dispatch callable.
Webui became a bit more complex and is in charge of selecting the
dispatch method, using dispatch for dispatchable parsers and passing
user defined or according to it's own logic.

Signed-off-by: Nir Izraeli <nir@sentinel-labs.com>