Form field names are now prefixed ids
Changed form input fields from parameter's name to full prefixed id. that complicated translation a bit in the POST function but lets us support two parameters with the same name and better future support of multiple subparsers.

Made a get_input() method so this type of changes will be easier later on.

Signed-off-by: Nir Izraeli <nir@sentinel-labs.com>